+++ Jump_Bruteforcer/PhysicsParams.cs
@@ -2,7 +2,7 @@
 {
     public static class PhysicsParams
     {
-        public const double GRAVITY = 0.4, MAX_VSPEED = 9, RELEASE_MULTIPLIER = 0.45, SJUMP_VSPEED = -8.5, DJUMP_VSPEED = -7;
+        public const float GRAVITY = 0.4f, MAX_VSPEED = 9, RELEASE_MULTIPLIER = 0.45f, SJUMP_VSPEED = -8.5f, DJUMP_VSPEED = -7;
         public const int WALKING_SPEED = 3;
     }
 }
